added cloud mysql database

+++ Client/src/App.js
@@ -8,25 +8,33 @@ function App() {
 
   let [textareaContent, setTextareaContent] = useState("")
   const [update, setUpdate] = useState(true)
+  const [allNotes, setAllNotes] = useState([])
+
+
+  /*Function to create a note to database with validations and with title and time......*/
 
   const addNote = () => {
-    if (!textareaContent) { return alert("Please write note") }
-    if(textareaContent.length>500){ return alert("You can write maximun 500 words") }
-    let options = {}
 
-    //Taking a title form all texts of note
+    if (!textareaContent) { return alert("Please write note") }
+    if (textareaContent.length > 500) { return alert("You can write maximun 500 words") }
+
+    let options = {}              //To store title description and time
+
+    // If title given by user the it take that word as a title 
+    // but if title is not given by the user then it assume first word as a title
 
     let arrayOfDes = textareaContent.split("\n")
 
-    if (arrayOfDes.length > 1 && arrayOfDes[0].length <= 20) {
+    if (arrayOfDes.length > 1 && arrayOfDes[0].length <= 20) {          //if title givem
 
       const [title, ...description] = arrayOfDes
+
       if (!description) { return alert("Write minimun two words to create a note...") }
 
       options.title = title
       options.description = description.join(" ")
 
-    } else {
+    } else {                                                            //if title not given
 
       let array = textareaContent.split(" ")
       const [title, ...description] = array
@@ -38,17 +46,16 @@ function App() {
       } else {
         return alert("First world should be assumed as a title so it's maximum length should be 20 !")
       }
-      options.description = description.join(" ")
+      options.description = textareaContent
     }
+
     let time = new Date()
 
     let currTime = `${time.getFullYear()}-${time.getMonth()}-${time.getDate()} ${time.getHours()}:${time.getMinutes()}:${time.getSeconds()}`
 
     options.createdAt = currTime;
 
-    console.log(options)
-
-    axios.post("http://localhost:5000/notes", options)
+    axios.post("http://localhost:5000/notes", options)               //create note request 
       .then(() => {
         alert("Data created succesfully")
         setTextareaContent("")
@@ -58,38 +65,37 @@ function App() {
   }
 
 
-  const [allNotes, setAllNotes] = useState([])
-
-
+  /*Function to read data (notes) from database....*/
 
   useEffect(() => {
     setUpdate(true)
     axios.get("http://localhost:5000/notes")
-      .then((res) =>{ console.log(res.data);setAllNotes(res.data)})
+      .then((res) => { setAllNotes(res.data.data) })
       .catch(err => console.log(err.message))
   }, [update])
 
 
-
+  /*Function to delete a notes from database....*/
 
   const deleteNotes = (i) => {
-    axios.delete(`http://localhost:5000/notes/${i}`)
-      .then(() => {
-        alert("Delete Succesfully")
-        setUpdate(false)
-      })
-      .catch(err => console.log(err.message))
+    let bool = window.confirm("Are you confirm to delete")
+    if (bool) {
+      axios.delete(`http://localhost:5000/notes/${i}`)
+        .then(() => { setUpdate(false) })
+        .catch(err => console.log(err.message))
+    }
   }
+
 
   return (
     <>
       <div id="navbar">
-        <span>Notes</span>
+        <span onClick={() => window.location.reload()}>Notes</span>
       </div>
 
       <div id="inputBox">
         <textarea value={textareaContent} placeholder='Take a note...' onChange={(e) => setTextareaContent(e.target.value)} />
-        <button onClick={addNote}>Add</button>
+        <button onClick={addNote} >Add</button>
       </div>
 
       <div id="allNotesBigBox">
